chore(test-loop): split receipt_to_tx tests into a module dir
Moves the 2102-line `test-loop-tests/src/tests/receipt_to_tx.rs` file
into a module directory:

  test-loop-tests/src/tests/receipt_to_tx/
  ├── mod.rs        (shared helpers + imports)
  ├── budget.rs     (per-request outcome budget)
  ├── classifier.rs (two-column tx/receipt classifier)
  ├── column.rs     (column-only path tests)
  ├── errors.rs     (WindowTooLarge, MalformedHint, OutcomesNotStored,
  │                 DepthExceeded, unresolvable-height)
  └── hint.rs       (hint-fallback walks, multi-hop, boundary refresh)

Pure refactor: no test logic changes, no new tests. The 37 existing
tests all keep their original bodies and assertions, only their file
location changes. `tests::receipt_to_tx` still resolves to the same
module path so existing `cargo test` filters keep working.
