+++ test-loop-tests/src/tests/receipt_to_tx/hint.rs
@@ -1,0 +1,875 @@
+//! Hint-fallback walks: tests that exercise the optional
+//! `(block_height, shard_id, window)` hint parameters of
+//! `EXPERIMENTAL_receipt_to_tx`. The column path remains the source of truth
+//! when the column is populated; the hint scan is the fallback.
+
+use super::*;
+
+/// `save_receipt_to_tx=false`, hint pointed at the tx execution height →
+/// terminal tx returned. The single-hop column miss falls back to the scan
+/// and walks the same outcome locally.
+#[test]
+fn test_hint_fallback_resolves_tx_origin() {
+    init_test_logger();
+    let user_account = create_account_id("account0");
+    let mut env = TestLoopBuilder::new()
+        .add_user_account(&user_account, Balance::from_near(1_000_000))
+        .epoch_length(EPOCH_LENGTH)
+        .track_all_shards()
+        .config_modifier(|config, _| {
+            config.save_receipt_to_tx = false;
+        })
+        .build();
+
+    let (tx_hash, receipt_id, height) = send_self_money(&mut env, &user_account, 1);
+
+    let response = handle(
+        &mut env,
+        GetReceiptToTx {
+            receipt_id,
+            block_height: Some(height),
+            shard_id: Some(ShardId::new(0)),
+            window: None,
+        },
+    )
+    .expect("hint should resolve to tx origin");
+    assert_eq!(response.transaction_hash, tx_hash);
+    assert_eq!(response.sender_account_id, user_account);
+}
+
+/// `save_receipt_to_tx=false`, height-only hint in a 2-shard setup. The
+/// handler does not know the creating shard, so hop 1 enumerates all shards
+/// at the hinted height and finds the transaction outcome.
+///
+/// Gated off under spice: the spice execution model places the cross-shard
+/// transaction outcome on a different block than the receipt's hinted height,
+/// so the scan window misses.
+#[test]
+#[cfg_attr(feature = "protocol_feature_spice", ignore)]
+fn test_hint_height_only_resolves_all_shards() {
+    init_test_logger();
+    let sender_account = create_account_id("account0");
+    let receiver_account: AccountId = "test1".parse().unwrap();
+    let mut env = TestLoopBuilder::new()
+        .num_shards(2)
+        .add_user_account(&sender_account, Balance::from_near(1_000_000))
+        .add_user_account(&receiver_account, Balance::from_near(1_000_000))
+        .epoch_length(EPOCH_LENGTH)
+        .track_all_shards()
+        .config_modifier(|config, _| {
+            config.save_receipt_to_tx = false;
+        })
+        .build();
+
+    let signer = create_user_test_signer(&sender_account);
+    let tx = SignedTransaction::send_money(
+        1,
+        sender_account.clone(),
+        receiver_account,
+        &signer,
+        Balance::from_yoctonear(100),
+        env.validator().head().last_block_hash,
+    );
+    let tx_hash = tx.get_hash();
+    let outcome = env.validator_runner().execute_tx(tx, Duration::seconds(10)).unwrap();
+    let receipt_id = outcome.transaction_outcome.outcome.receipt_ids[0];
+    let tx_height = env
+        .validator()
+        .client()
+        .chain
+        .get_block_header(&outcome.transaction_outcome.block_hash)
+        .unwrap()
+        .height();
+
+    let response = handle(
+        &mut env,
+        GetReceiptToTx { receipt_id, block_height: Some(tx_height), shard_id: None, window: None },
+    )
+    .expect("height-only hint should scan all shards and resolve to tx origin");
+    assert_eq!(response.transaction_hash, tx_hash);
+    assert_eq!(response.sender_account_id, sender_account);
+}
+
+/// `save_receipt_to_tx=false`, contract refund chain (depth 2). Hint at the
+/// action receipt's execution height. The handler walks both hops server-side
+/// via repeated hint scans.
+///
+/// Gated off under spice: the spice execution model produces the refund and
+/// action receipts on different blocks than the standard model, so the
+/// computed hint coordinates don't match the outcome rows the scan inspects.
+#[test]
+#[cfg_attr(feature = "protocol_feature_spice", ignore)]
+fn test_hint_fallback_resolves_through_refund_chain() {
+    init_test_logger();
+    let user_account = create_account_id("account0");
+    let min_gas_price = Balance::from_yoctonear(100_000_000);
+    let mut env = TestLoopBuilder::new()
+        .add_user_account(&user_account, Balance::from_near(1_000_000))
+        .epoch_length(EPOCH_LENGTH)
+        .track_all_shards()
+        .gas_prices(min_gas_price, min_gas_price)
+        .config_modifier(|config, _| {
+            config.save_receipt_to_tx = false;
+        })
+        .build();
+
+    let signer = create_user_test_signer(&user_account);
+    let deploy_tx = SignedTransaction::deploy_contract(
+        1,
+        &user_account,
+        near_test_contracts::rs_contract().to_vec(),
+        &signer,
+        env.validator().head().last_block_hash,
+    );
+    env.validator_runner().run_tx(deploy_tx, Duration::seconds(5));
+
+    let call_tx = SignedTransaction::call(
+        2,
+        user_account.clone(),
+        user_account.clone(),
+        &signer,
+        Balance::ZERO,
+        "log_something".to_owned(),
+        vec![],
+        Gas::from_teragas(300),
+        env.validator().head().last_block_hash,
+    );
+    let call_tx_hash = call_tx.get_hash();
+    let outcome = env.validator_runner().execute_tx(call_tx, Duration::seconds(10)).unwrap();
+    let action_receipt_id = outcome.transaction_outcome.outcome.receipt_ids[0];
+    let action_outcome =
+        outcome.receipts_outcome.iter().find(|r| r.id == action_receipt_id).unwrap();
+    let refund_receipt_id = action_outcome.outcome.receipt_ids[0];
+    let action_height = env
+        .validator()
+        .client()
+        .chain
+        .get_block_header(&action_outcome.block_hash)
+        .unwrap()
+        .height();
+
+    let response = handle(
+        &mut env,
+        GetReceiptToTx {
+            receipt_id: refund_receipt_id,
+            block_height: Some(action_height),
+            shard_id: Some(ShardId::new(0)),
+            window: None,
+        },
+    )
+    .expect("hint walk should resolve refund → action receipt → tx");
+    assert_eq!(response.transaction_hash, call_tx_hash);
+    assert_eq!(response.sender_account_id, user_account);
+}
+
+/// Cross-shard depth-2 walk with `save_receipt_to_tx=false`. Hop 1 uses the
+/// supplied action shard to resolve refund → action receipt. The hop-2 scan
+/// is shard-narrowed via the handler's predecessor-account derivation: the
+/// action receipt's `parent_predecessor_id` resolves to the sender shard, so
+/// the ancestor scan goes straight to shard 0 and finds the originating
+/// transaction without enumerating all shards.
+///
+/// Gated off under spice: the spice execution model lands the cross-shard
+/// refund / action receipts on different blocks than the standard model, so
+/// the hint coordinates this test computes don't line up with the outcome
+/// rows that the resolver scans.
+#[test]
+#[cfg_attr(feature = "protocol_feature_spice", ignore)]
+fn test_hint_cross_shard_walk_resolves_via_predecessor_shard() {
+    init_test_logger();
+    let sender_account = create_account_id("account0");
+    let receiver_account: AccountId = "test1".parse().unwrap();
+    let min_gas_price = Balance::from_yoctonear(100_000_000);
+    let mut env = TestLoopBuilder::new()
+        .num_shards(2)
+        .add_user_account(&sender_account, Balance::from_near(1_000_000))
+        .add_user_account(&receiver_account, Balance::from_near(1_000_000))
+        .epoch_length(EPOCH_LENGTH)
+        .track_all_shards()
+        .gas_prices(min_gas_price, min_gas_price)
+        .config_modifier(|config, _| {
+            config.save_receipt_to_tx = false;
+        })
+        .build();
+
+    let receiver_signer = create_user_test_signer(&receiver_account);
+    let deploy_tx = SignedTransaction::deploy_contract(
+        1,
+        &receiver_account,
+        near_test_contracts::rs_contract().to_vec(),
+        &receiver_signer,
+        env.validator().head().last_block_hash,
+    );
+    env.validator_runner().run_tx(deploy_tx, Duration::seconds(5));
+
+    let sender_signer = create_user_test_signer(&sender_account);
+    let call_tx = SignedTransaction::call(
+        1,
+        sender_account.clone(),
+        receiver_account,
+        &sender_signer,
+        Balance::ZERO,
+        "log_something".to_owned(),
+        vec![],
+        Gas::from_teragas(300),
+        env.validator().head().last_block_hash,
+    );
+    let call_tx_hash = call_tx.get_hash();
+    let outcome = env.validator_runner().execute_tx(call_tx, Duration::seconds(10)).unwrap();
+    let action_receipt_id = outcome.transaction_outcome.outcome.receipt_ids[0];
+    let action_outcome =
+        outcome.receipts_outcome.iter().find(|r| r.id == action_receipt_id).unwrap();
+    let refund_receipt_id = action_outcome.outcome.receipt_ids[0];
+    let action_height = env
+        .validator()
+        .client()
+        .chain
+        .get_block_header(&action_outcome.block_hash)
+        .unwrap()
+        .height();
+    let action_shard = shard_containing_outcome(
+        &env,
+        action_height,
+        action_receipt_id,
+        &[ShardId::new(0), ShardId::new(1)],
+    );
+
+    let response = handle(
+        &mut env,
+        GetReceiptToTx {
+            receipt_id: refund_receipt_id,
+            block_height: Some(action_height),
+            shard_id: Some(action_shard),
+            window: None,
+        },
+    )
+    .expect("cross-shard refund chain should resolve via all-shards ancestor scan");
+    assert_eq!(response.transaction_hash, call_tx_hash);
+    assert_eq!(response.sender_account_id, sender_account);
+}
+
+/// `save_receipt_to_tx=true`, `save_tx_outcomes=false`, hint supplied but
+/// never used because the column resolves every hop. Must succeed —
+/// `OutcomesNotStored` should only fire when a scan is actually needed.
+#[test]
+fn test_hint_with_column_populated_save_tx_outcomes_false_succeeds() {
+    init_test_logger();
+    let user_account = create_account_id("account0");
+    let mut env = TestLoopBuilder::new()
+        .add_user_account(&user_account, Balance::from_near(1_000_000))
+        .epoch_length(EPOCH_LENGTH)
+        .track_all_shards()
+        .config_modifier(|config, _| {
+            config.save_tx_outcomes = false;
+        })
+        .build();
+
+    let signer = create_user_test_signer(&user_account);
+    let tx = SignedTransaction::send_money(
+        1,
+        user_account.clone(),
+        user_account,
+        &signer,
+        Balance::from_yoctonear(100),
+        env.validator().head().last_block_hash,
+    );
+    let tx_hash = tx.get_hash();
+    env.validator().submit_tx(tx);
+    let target_height = env.validator().head().height + 2 * EPOCH_LENGTH;
+    env.validator_runner().run_until_executed_height(target_height);
+
+    // Find the receipt_id from the ReceiptToTx column directly (outcomes are not stored).
+    let store = env.validator().store();
+    let (receipt_id, _) = store
+        .iter_ser::<ReceiptToTxInfo>(DBCol::ReceiptToTx)
+        .find_map(|(k, info)| match &info {
+            ReceiptToTxInfo::V1(v1) => match &v1.origin {
+                ReceiptOrigin::FromTransaction(o) if o.tx_hash == tx_hash => {
+                    Some((CryptoHash::try_from(k.as_ref()).unwrap(), info))
+                }
+                _ => None,
+            },
+        })
+        .expect("column entry should exist");
+
+    // Supply a hint that, if it triggered, would hit OutcomesNotStored. The
+    // column path should answer first and ignore the hint.
+    let response = handle(
+        &mut env,
+        GetReceiptToTx {
+            receipt_id,
+            block_height: Some(1),
+            shard_id: Some(ShardId::new(0)),
+            window: None,
+        },
+    )
+    .expect("column hit must short-circuit OutcomesNotStored");
+    assert_eq!(response.transaction_hash, tx_hash);
+}
+
+/// Hint far outside the receipt's window → `UnknownReceipt` (no column entry,
+/// scan window exhausted).
+#[test]
+fn test_hint_fallback_wrong_height() {
+    init_test_logger();
+    let user_account = create_account_id("account0");
+    let mut env = TestLoopBuilder::new()
+        .add_user_account(&user_account, Balance::from_near(1_000_000))
+        .epoch_length(EPOCH_LENGTH)
+        .track_all_shards()
+        .config_modifier(|config, _| {
+            config.save_receipt_to_tx = false;
+        })
+        .build();
+
+    let (_, receipt_id, height) = send_self_money(&mut env, &user_account, 1);
+    let bogus_hint = height + 100;
+    let result = handle(
+        &mut env,
+        GetReceiptToTx {
+            receipt_id,
+            block_height: Some(bogus_hint),
+            shard_id: Some(ShardId::new(0)),
+            window: None,
+        },
+    );
+    match result {
+        Err(GetReceiptToTxError::UnknownReceipt(id)) => {
+            assert_eq!(id, receipt_id, "should report the queried receipt that couldn't be found");
+        }
+        other => panic!("expected UnknownReceipt, got {other:?}"),
+    }
+}
+
+/// Synthetic chain: column has child → FromReceipt(P), but the column entry
+/// for P is absent. Hint supplied. The next iteration's column-miss scan
+/// should pick up P's coordinates and resolve terminally. Regression guard
+/// for mixed column-hit / hint-fallback walks.
+#[test]
+fn test_hint_column_then_fallback_boundary() {
+    init_test_logger();
+    let user_account = create_account_id("account0");
+    let min_gas_price = Balance::from_yoctonear(100_000_000);
+    let mut env = TestLoopBuilder::new()
+        .add_user_account(&user_account, Balance::from_near(1_000_000))
+        .epoch_length(EPOCH_LENGTH)
+        .track_all_shards()
+        .gas_prices(min_gas_price, min_gas_price)
+        .build();
+
+    let signer = create_user_test_signer(&user_account);
+    let deploy_tx = SignedTransaction::deploy_contract(
+        1,
+        &user_account,
+        near_test_contracts::rs_contract().to_vec(),
+        &signer,
+        env.validator().head().last_block_hash,
+    );
+    env.validator_runner().run_tx(deploy_tx, Duration::seconds(5));
+
+    let call_tx = SignedTransaction::call(
+        2,
+        user_account.clone(),
+        user_account.clone(),
+        &signer,
+        Balance::ZERO,
+        "log_something".to_owned(),
+        vec![],
+        Gas::from_teragas(300),
+        env.validator().head().last_block_hash,
+    );
+    let call_tx_hash = call_tx.get_hash();
+    let outcome = env.validator_runner().execute_tx(call_tx, Duration::seconds(10)).unwrap();
+    let action_receipt_id = outcome.transaction_outcome.outcome.receipt_ids[0];
+    let action_outcome =
+        outcome.receipts_outcome.iter().find(|r| r.id == action_receipt_id).unwrap();
+    let refund_receipt_id = action_outcome.outcome.receipt_ids[0];
+    let action_height = env
+        .validator()
+        .client()
+        .chain
+        .get_block_header(&action_outcome.block_hash)
+        .unwrap()
+        .height();
+
+    // Delete the column entry for the parent (action_receipt_id) so the next
+    // iteration column-misses and must fall back to the hint scan.
+    let store = env.validator().store();
+    let mut store_update = store.store_update();
+    store_update.delete(DBCol::ReceiptToTx, action_receipt_id.as_ref());
+    store_update.commit();
+    assert!(
+        store.get_ser::<ReceiptToTxInfo>(DBCol::ReceiptToTx, action_receipt_id.as_ref()).is_none(),
+        "test setup: action receipt's column entry must be gone"
+    );
+
+    let response = handle(
+        &mut env,
+        GetReceiptToTx {
+            receipt_id: refund_receipt_id,
+            block_height: Some(action_height),
+            shard_id: Some(ShardId::new(0)),
+            window: None,
+        },
+    )
+    .expect("column hit then next-hop column-miss scan should resolve to terminal tx");
+    assert_eq!(response.transaction_hash, call_tx_hash);
+    assert_eq!(response.sender_account_id, user_account);
+}
+
+/// Cross-shard hint walk: 2-shard setup, `save_receipt_to_tx=false`,
+/// cross-shard transfer (sender on shard 0, receiver on shard 1).
+///
+/// The action receipt produced by the tx executes on the *receiver's* shard;
+/// the tx itself executes on the *sender's* shard. A hint pointing at the
+/// action receipt's execution shard cannot find the tx outcome on the
+/// originating shard — the center-out scan walks the wrong shard's
+/// `OutcomeIds` rows.
+///
+/// We hint exactly at the action receipt's execution coordinates. The hint
+/// scan misses (the tx outcome is on the other shard), no column entry
+/// exists, and the handler returns `UnknownReceipt` at the cross-shard
+/// boundary rather than fabricating a result.
+///
+/// Regression guard for the documented best-effort failure mode. If a
+/// future change adds shard-aware hint derivation (e.g. via
+/// `parent.predecessor_id() -> account_id_to_shard_id`), this test should
+/// be updated to reflect the new contract.
+#[test]
+fn test_hint_fallback_cross_shard_returns_unknown_receipt() {
+    init_test_logger();
+    let sender_account = create_account_id("account0");
+    let receiver_account: AccountId = "test1".parse().unwrap();
+
+    let mut env = TestLoopBuilder::new()
+        .num_shards(2)
+        .add_user_account(&sender_account, Balance::from_near(1_000_000))
+        .add_user_account(&receiver_account, Balance::from_near(1_000_000))
+        .epoch_length(EPOCH_LENGTH)
+        .track_all_shards()
+        .config_modifier(|config, _| {
+            config.save_receipt_to_tx = false;
+        })
+        .build();
+
+    let signer = create_user_test_signer(&sender_account);
+    let tx = SignedTransaction::send_money(
+        1,
+        sender_account,
+        receiver_account,
+        &signer,
+        Balance::from_yoctonear(100),
+        env.validator().head().last_block_hash,
+    );
+    let outcome = env.validator_runner().execute_tx(tx, Duration::seconds(10)).unwrap();
+
+    let action_receipt_id = outcome.transaction_outcome.outcome.receipt_ids[0];
+    let action_outcome = outcome
+        .receipts_outcome
+        .iter()
+        .find(|r| r.id == action_receipt_id)
+        .expect("action receipt outcome should exist");
+    let action_height = env
+        .validator()
+        .client()
+        .chain
+        .get_block_header(&action_outcome.block_hash)
+        .unwrap()
+        .height();
+
+    // Identify which shard the action receipt actually executed on. We don't
+    // hardcode the layout — multi_shard(2) puts the boundary at "test1" but a
+    // future layout change should still leave this test exercising the
+    // cross-shard scenario.
+    let action_shard = shard_containing_outcome(
+        &env,
+        action_height,
+        action_receipt_id,
+        &[ShardId::new(0), ShardId::new(1)],
+    );
+
+    // Query the action receipt itself with the hint pointed at *its* execution
+    // shard. The tx outcome that produced this receipt lives on the *other*
+    // shard, so the hint scan walks the wrong shard's outcomes and misses.
+    let handle = env.node_datas[0].view_client_sender.actor_handle();
+    let view_client: &mut near_client::ViewClientActor = env.test_loop.data.get_mut(&handle);
+    let result = view_client.handle(GetReceiptToTx {
+        receipt_id: action_receipt_id,
+        block_height: Some(action_height),
+        shard_id: Some(action_shard),
+        window: None,
+    });
+
+    match result {
+        Err(GetReceiptToTxError::UnknownReceipt(id)) => {
+            assert_eq!(
+                id, action_receipt_id,
+                "expected UnknownReceipt for the action receipt; its producing tx is on the other shard"
+            );
+        }
+        other => panic!(
+            "expected UnknownReceipt at the cross-shard hop; got {other:?}. \
+             If a future change adds shard-aware hint advancement, update this test."
+        ),
+    }
+}
+
+/// Stale-hint fall-through: the next-hop column-miss scan misses AND no
+/// later column entry hits, so the walk falls through to `UnknownReceipt`
+/// rather than fabricating a result.
+///
+/// The sibling scenario (refresh misses, column hits, terminal Ok) is
+/// already covered by `test_hint_column_then_fallback_boundary`.
+#[test]
+fn test_hint_stale_then_column_miss_returns_unknown() {
+    init_test_logger();
+    let user_account = create_account_id("account0");
+    let min_gas_price = Balance::from_yoctonear(100_000_000);
+    let mut env = TestLoopBuilder::new()
+        .add_user_account(&user_account, Balance::from_near(1_000_000))
+        .epoch_length(EPOCH_LENGTH)
+        .track_all_shards()
+        .gas_prices(min_gas_price, min_gas_price)
+        .build();
+
+    let signer = create_user_test_signer(&user_account);
+    let deploy_tx = SignedTransaction::deploy_contract(
+        1,
+        &user_account,
+        near_test_contracts::rs_contract().to_vec(),
+        &signer,
+        env.validator().head().last_block_hash,
+    );
+    env.validator_runner().run_tx(deploy_tx, Duration::seconds(5));
+
+    let call_tx = SignedTransaction::call(
+        2,
+        user_account.clone(),
+        user_account,
+        &signer,
+        Balance::ZERO,
+        "log_something".to_owned(),
+        vec![],
+        Gas::from_teragas(300),
+        env.validator().head().last_block_hash,
+    );
+    let outcome = env.validator_runner().execute_tx(call_tx, Duration::seconds(10)).unwrap();
+    let action_receipt_id = outcome.transaction_outcome.outcome.receipt_ids[0];
+    let action_outcome =
+        outcome.receipts_outcome.iter().find(|r| r.id == action_receipt_id).unwrap();
+    let refund_receipt_id = action_outcome.outcome.receipt_ids[0];
+    let action_height = env
+        .validator()
+        .client()
+        .chain
+        .get_block_header(&action_outcome.block_hash)
+        .unwrap()
+        .height();
+
+    // Delete the column entry for action_receipt_id (the parent the walk
+    // recurses on). With save_receipt_to_tx still on, hop 1 column-hits for
+    // refund_receipt_id and returns FromReceipt(action_receipt_id). Hop 2
+    // column-misses for the parent; with a wildly stale hint, the scan also
+    // misses and the walk surfaces UnknownReceipt.
+    let store = env.validator().store();
+    let mut store_update = store.store_update();
+    store_update.delete(DBCol::ReceiptToTx, action_receipt_id.as_ref());
+    store_update.commit();
+
+    let stale_height = action_height + 10 * EPOCH_LENGTH;
+    let result = handle(
+        &mut env,
+        GetReceiptToTx {
+            receipt_id: refund_receipt_id,
+            block_height: Some(stale_height),
+            shard_id: Some(ShardId::new(0)),
+            window: Some(2),
+        },
+    );
+    match result {
+        Err(GetReceiptToTxError::UnknownReceipt(id)) => {
+            assert_eq!(id, action_receipt_id);
+        }
+        other => {
+            panic!("expected UnknownReceipt at the stale-hint cross-hop miss, got {other:?}")
+        }
+    }
+}
+
+/// 2-hop self-call walk with the ancestor scan distance pinned to 0. The hint
+/// resolver visits only the anchor height for hop 1+, so the test only passes
+/// if the producing outcome of the parent receipt lives at the anchor itself.
+/// For a same-account self-call, `process_local_receipts` runs the action
+/// receipt within the same `apply()` call as its emitting tx, so both
+/// outcomes share a block. Locks the anchor-inclusion invariant in the
+/// integration layer.
+#[test]
+#[cfg_attr(feature = "protocol_feature_spice", ignore)]
+fn test_hint_ancestor_includes_anchor() {
+    init_test_logger();
+    let user_account = create_account_id("account0");
+    let min_gas_price = Balance::from_yoctonear(100_000_000);
+    let mut env = TestLoopBuilder::new()
+        .add_user_account(&user_account, Balance::from_near(1_000_000))
+        .epoch_length(EPOCH_LENGTH)
+        .track_all_shards()
+        .gas_prices(min_gas_price, min_gas_price)
+        .config_modifier(|config, _| {
+            config.save_receipt_to_tx = false;
+            // Anchor-only: hop 1+ may only inspect the resolved parent's
+            // own execution block. A successful walk proves the anchor is
+            // visited.
+            config.receipt_to_tx_max_hop_distance = 0;
+        })
+        .build();
+
+    let signer = create_user_test_signer(&user_account);
+    let deploy_tx = SignedTransaction::deploy_contract(
+        1,
+        &user_account,
+        near_test_contracts::rs_contract().to_vec(),
+        &signer,
+        env.validator().head().last_block_hash,
+    );
+    env.validator_runner().run_tx(deploy_tx, Duration::seconds(5));
+
+    let call_tx = SignedTransaction::call(
+        2,
+        user_account.clone(),
+        user_account.clone(),
+        &signer,
+        Balance::ZERO,
+        "log_something".to_owned(),
+        vec![],
+        Gas::from_teragas(300),
+        env.validator().head().last_block_hash,
+    );
+    let call_tx_hash = call_tx.get_hash();
+    let outcome = env.validator_runner().execute_tx(call_tx, Duration::seconds(10)).unwrap();
+    let action_receipt_id = outcome.transaction_outcome.outcome.receipt_ids[0];
+    let action_outcome =
+        outcome.receipts_outcome.iter().find(|r| r.id == action_receipt_id).unwrap();
+    let refund_receipt_id = action_outcome.outcome.receipt_ids[0];
+    let action_height = env
+        .validator()
+        .client()
+        .chain
+        .get_block_header(&action_outcome.block_hash)
+        .unwrap()
+        .height();
+
+    let response = handle(
+        &mut env,
+        GetReceiptToTx {
+            receipt_id: refund_receipt_id,
+            block_height: Some(action_height),
+            shard_id: Some(ShardId::new(0)),
+            window: None,
+        },
+    )
+    .expect("anchor-only ancestor scan should still resolve local same-shard receipts");
+    assert_eq!(response.transaction_hash, call_tx_hash);
+    assert_eq!(response.sender_account_id, user_account);
+}
+
+/// Inject an emit→execute delay larger than `receipt_to_tx_max_hop_distance`
+/// via a cross-shard transfer. With `max_hop_distance=0` and the caller's
+/// `window=0`, the next-iteration column-miss scan in Ancestor mode cannot
+/// reach the tx's producing block. The walk surfaces
+/// `UnknownReceipt`, documenting the fail-fast contract: ancestor misses
+/// when configured distance is too tight for the actual emit-to-execute
+/// delay.
+#[test]
+#[cfg_attr(feature = "protocol_feature_spice", ignore)]
+fn test_hint_ancestor_distance_misses_when_delay_exceeds_config() {
+    init_test_logger();
+    let sender_account = create_account_id("account0");
+    let receiver_account: AccountId = "test1".parse().unwrap();
+    let min_gas_price = Balance::from_yoctonear(100_000_000);
+    let mut env = TestLoopBuilder::new()
+        .num_shards(2)
+        .add_user_account(&sender_account, Balance::from_near(1_000_000))
+        .add_user_account(&receiver_account, Balance::from_near(1_000_000))
+        .epoch_length(EPOCH_LENGTH)
+        .track_all_shards()
+        .gas_prices(min_gas_price, min_gas_price)
+        .config_modifier(|config, _| {
+            config.save_receipt_to_tx = false;
+            // Distance 0 collapses the ancestor scan to anchor-only.
+            config.receipt_to_tx_max_hop_distance = 0;
+        })
+        .build();
+
+    let signer = create_user_test_signer(&sender_account);
+    let tx = SignedTransaction::send_money(
+        1,
+        sender_account,
+        receiver_account,
+        &signer,
+        Balance::from_yoctonear(100),
+        env.validator().head().last_block_hash,
+    );
+    let outcome = env.validator_runner().execute_tx(tx, Duration::seconds(10)).unwrap();
+    let action_receipt_id = outcome.transaction_outcome.outcome.receipt_ids[0];
+    let action_outcome =
+        outcome.receipts_outcome.iter().find(|r| r.id == action_receipt_id).unwrap();
+    let action_height = env
+        .validator()
+        .client()
+        .chain
+        .get_block_header(&action_outcome.block_hash)
+        .unwrap()
+        .height();
+    let action_shard = shard_containing_outcome(
+        &env,
+        action_height,
+        action_receipt_id,
+        &[ShardId::new(0), ShardId::new(1)],
+    );
+
+    // window=0 forces hop 0 to inspect only the anchor height too; combined
+    // with max_hop_distance=0 the walk has zero slack to reach the
+    // cross-shard tx outcome which executed a block earlier.
+    let result = handle(
+        &mut env,
+        GetReceiptToTx {
+            receipt_id: action_receipt_id,
+            block_height: Some(action_height),
+            shard_id: Some(action_shard),
+            window: Some(0),
+        },
+    );
+    match result {
+        Err(GetReceiptToTxError::UnknownReceipt(id)) => {
+            assert_eq!(id, action_receipt_id);
+        }
+        other => panic!("expected UnknownReceipt under tight max_hop_distance, got {other:?}"),
+    }
+}
+
+/// Operator override: raise `receipt_to_tx_max_hint_window` past the default
+/// and verify that a `window` value over the old default is accepted instead
+/// of being rejected by the up-front cap check. Locks the contract that the
+/// new field is operator-tunable.
+#[test]
+fn test_hint_window_config_override() {
+    init_test_logger();
+    let mut env = TestLoopBuilder::new()
+        .epoch_length(EPOCH_LENGTH)
+        .track_all_shards()
+        .config_modifier(|config, _| {
+            config.receipt_to_tx_max_hint_window = 50;
+        })
+        .build();
+
+    // 30 > default cap (20) but < operator-raised cap (50). The handler must
+    // accept it; with no matching receipt anywhere the walk terminates in
+    // UnknownReceipt, not WindowTooLarge.
+    let result = handle(
+        &mut env,
+        GetReceiptToTx {
+            receipt_id: CryptoHash::hash_bytes(b"absent"),
+            block_height: Some(100),
+            shard_id: Some(ShardId::new(0)),
+            window: Some(30),
+        },
+    );
+    assert!(
+        matches!(result, Err(GetReceiptToTxError::UnknownReceipt(_))),
+        "window=30 must be accepted under max_hint_window=50; got {result:?}"
+    );
+}
+
+/// Pin the load-bearing band for hop-1+ ancestor scan width:
+/// `d ∈ (effective_window, max_hop_distance]` where
+/// `d = parent_execution_height − grandparent_execution_height`.
+///
+/// Setup mirrors `test_hint_cross_shard_walk_resolves_via_predecessor_shard`
+/// (cross-shard contract call producing a refund chain) but tightens the
+/// caller's window to 0 so `effective_window = 0 < d ≤ max_hop_distance = 20`.
+/// The natural cross-shard delay puts the producing tx outcome a block earlier
+/// than the action receipt on the sender shard.
+///
+/// Hop 0 finds the action receipt at `(action_height, action_shard)` with
+/// `window = 0` (anchor only). Without the hop-1+ ancestor scan width
+/// (`max_hop_distance = 20`) covering the gap between the action receipt and
+/// its producing tx, the walk would terminate `UnknownReceipt`. This test pins
+/// the load-bearing claim so any future restructure that preserves the same
+/// coverage continues to pass.
+#[test]
+#[cfg_attr(feature = "protocol_feature_spice", ignore)]
+fn test_hint_ancestor_gap_band() {
+    init_test_logger();
+    let sender_account = create_account_id("account0");
+    let receiver_account: AccountId = "test1".parse().unwrap();
+    let min_gas_price = Balance::from_yoctonear(100_000_000);
+    let mut env = TestLoopBuilder::new()
+        .num_shards(2)
+        .add_user_account(&sender_account, Balance::from_near(1_000_000))
+        .add_user_account(&receiver_account, Balance::from_near(1_000_000))
+        .epoch_length(EPOCH_LENGTH)
+        .track_all_shards()
+        .gas_prices(min_gas_price, min_gas_price)
+        .config_modifier(|config, _| {
+            config.save_receipt_to_tx = false;
+        })
+        .build();
+
+    let receiver_signer = create_user_test_signer(&receiver_account);
+    let deploy_tx = SignedTransaction::deploy_contract(
+        1,
+        &receiver_account,
+        near_test_contracts::rs_contract().to_vec(),
+        &receiver_signer,
+        env.validator().head().last_block_hash,
+    );
+    env.validator_runner().run_tx(deploy_tx, Duration::seconds(5));
+
+    let sender_signer = create_user_test_signer(&sender_account);
+    let call_tx = SignedTransaction::call(
+        1,
+        sender_account.clone(),
+        receiver_account,
+        &sender_signer,
+        Balance::ZERO,
+        "log_something".to_owned(),
+        vec![],
+        Gas::from_teragas(300),
+        env.validator().head().last_block_hash,
+    );
+    let call_tx_hash = call_tx.get_hash();
+    let outcome = env.validator_runner().execute_tx(call_tx, Duration::seconds(10)).unwrap();
+    let action_receipt_id = outcome.transaction_outcome.outcome.receipt_ids[0];
+    let action_outcome =
+        outcome.receipts_outcome.iter().find(|r| r.id == action_receipt_id).unwrap();
+    let refund_receipt_id = action_outcome.outcome.receipt_ids[0];
+    let action_height = env
+        .validator()
+        .client()
+        .chain
+        .get_block_header(&action_outcome.block_hash)
+        .unwrap()
+        .height();
+    let action_shard = shard_containing_outcome(
+        &env,
+        action_height,
+        action_receipt_id,
+        &[ShardId::new(0), ShardId::new(1)],
+    );
+
+    // window=0 forces effective_window=0 < d on hop 1+. Walk only succeeds
+    // because the hop-1+ ancestor scan width (max_hop_distance default 20)
+    // reaches the producing tx outcome a block earlier on the sender shard.
+    let response = handle(
+        &mut env,
+        GetReceiptToTx {
+            receipt_id: refund_receipt_id,
+            block_height: Some(action_height),
+            shard_id: Some(action_shard),
+            window: Some(0),
+        },
+    )
+    .expect("ancestor gap band must be covered by hop-1+ scan width");
+    assert_eq!(response.transaction_hash, call_tx_hash);
+    assert_eq!(response.sender_account_id, sender_account);
+}
